third_party/blink/web_tests/inspector-protocol/css/css-set-multiple-style-texts-correct-ranges-expected.txt

Verifies that the range of i-th payload corresponds to the state at the end of i-th edit; not the state at the end of all edits in setStyleTexts operation.
==== Initial style sheet text ====
#test {
    box-sizing: border-box;
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}



Running test: removeTwoStyleTexts
StylePayload result ranges:
    range #0: {4, 7, 4, 7}
    range #1: {0, 7, 0, 7}
